move o1 detection from HIR to runtime
