+++ baml_language/crates/sys_llm/src/build_request/openai/chat_completions.rs
@@ -1,7 +1,6 @@
-//! OpenAI-format HTTP request builder.
+//! `OpenAI` Chat Completions request builder.
 //!
-//! Supports: `OpenAi`, `OpenAiGeneric`, `AzureOpenAi`, Ollama, `OpenRouter`,
-//! and `OpenAiResponses` (Responses API).
+//! Supports: `OpenAi`, `OpenAiGeneric`, `AzureOpenAi`, Ollama, `OpenRouter`.
 
 use std::fmt::Write;
 
@@ -9,8 +8,10 @@ use baml_builtins::{PromptAst, PromptAstSimple};
 use indexmap::IndexMap;
 use serde::Serialize;
 
-use super::{BuildRequestError, LlmPrimitiveClient, LlmRequestBuilder, get_string_option};
-use crate::LlmProvider;
+use crate::{
+    LlmProvider,
+    build_request::{BuildRequestError, LlmPrimitiveClient, LlmRequestBuilder, get_string_option},
+};
 
 /// A single chat message in the `OpenAI` Chat Completions format.
 #[derive(Serialize)]
@@ -325,231 +326,6 @@ fn openai_media_part(
         },
         MediaKind::Video => Err(BuildRequestError::UnsupportedMedia(
             "video input is not supported on OpenAI chat completions".into(),
-        )),
-    }
-}
-
-/// A single message in the `OpenAI` Responses API format.
-#[derive(Debug, Serialize)]
-struct ResponsesMessage {
-    role: String,
-    content: Vec<ResponsesContentPart>,
-}
-
-/// A content part within a Responses API message.
-///
-/// Uses `input_text`/`output_text` instead of just `text`, and `input_image`/`input_audio`/`input_file`
-/// instead of `image_url`/`input_audio`/`file`.
-#[derive(Debug, Serialize)]
-#[serde(tag = "type", rename_all = "snake_case")]
-enum ResponsesContentPart {
-    InputText {
-        text: String,
-    },
-    OutputText {
-        text: String,
-    },
-    InputImage {
-        #[serde(skip_serializing_if = "Option::is_none")]
-        detail: Option<String>,
-        image_url: String,
-    },
-    InputAudio {
-        input_audio: InputAudio,
-    },
-    InputFile {
-        #[serde(skip_serializing_if = "Option::is_none")]
-        file_url: Option<String>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        file_data: Option<String>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        file_id: Option<String>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        filename: Option<String>,
-    },
-}
-
-/// Builder for the `OpenAI` Responses API (`/v1/responses`).
-///
-/// Uses `"input"` instead of `"messages"` and different content part types:
-/// `input_text`, `output_text`, `input_image`, `input_audio`, `input_file`.
-pub(crate) struct OpenAiResponsesBuilder;
-
-impl LlmRequestBuilder for OpenAiResponsesBuilder {
-    fn provider_skip_keys(&self) -> &'static [&'static str] {
-        &[]
-    }
-
-    fn build_url(&self, client: &LlmPrimitiveClient) -> Result<String, BuildRequestError> {
-        let base_url = get_string_option(client, "base_url")
-            .unwrap_or_else(|| "https://api.openai.com/v1".to_string());
-        let base_url = base_url.trim_end_matches('/');
-        Ok(format!("{base_url}/responses"))
-    }
-
-    fn build_auth_headers(&self, client: &LlmPrimitiveClient) -> IndexMap<String, String> {
-        let mut headers = IndexMap::new();
-        if let Some(api_key) = get_string_option(client, "api_key") {
-            headers.insert("authorization".to_string(), format!("Bearer {api_key}"));
-        }
-        headers
-    }
-
-    fn build_prompt_body(
-        &self,
-        client: &LlmPrimitiveClient,
-        prompt: bex_vm_types::PromptAst,
-    ) -> Result<serde_json::Map<String, serde_json::Value>, BuildRequestError> {
-        let mut map = serde_json::Map::new();
-        let input = prompt_to_responses_input(&prompt, &client.default_role)?;
-        map.insert(
-            "input".to_string(),
-            serde_json::to_value(input).expect("infallible"),
-        );
-        Ok(map)
-    }
-}
-
-/// Converts a top-level [`PromptAst`] into a list of Responses API input messages.
-fn prompt_to_responses_input(
-    prompt: &bex_vm_types::PromptAst,
-    default_role: &str,
-) -> Result<Vec<ResponsesMessage>, BuildRequestError> {
-    match prompt.as_ref() {
-        PromptAst::Vec(items) => items
-            .iter()
-            .map(|node| responses_node_to_message(node, default_role))
-            .collect(),
-        _ => Ok(vec![responses_node_to_message(prompt, default_role)?]),
-    }
-}
-
-/// Converts a single [`PromptAst`] node into a Responses API input message.
-fn responses_node_to_message(
-    node: &bex_vm_types::PromptAst,
-    default_role: &str,
-) -> Result<ResponsesMessage, BuildRequestError> {
-    match node.as_ref() {
-        PromptAst::Message { role, content, .. } => {
-            let parts = responses_content_parts(content.as_ref(), role)?;
-            Ok(ResponsesMessage {
-                role: role.clone(),
-                content: parts,
-            })
-        }
-        PromptAst::Simple(content) => {
-            let parts = responses_content_parts(content.as_ref(), default_role)?;
-            Ok(ResponsesMessage {
-                role: default_role.to_string(),
-                content: parts,
-            })
-        }
-        PromptAst::Vec(_) => unreachable!("Nested vecs should not appear after specialization"),
-    }
-}
-
-/// Converts a [`PromptAstSimple`] content node into Responses API content parts.
-///
-/// Assistant-role text uses `output_text`; all other roles use `input_text`.
-fn responses_content_parts(
-    content: &PromptAstSimple,
-    role: &str,
-) -> Result<Vec<ResponsesContentPart>, BuildRequestError> {
-    match content {
-        PromptAstSimple::String(s) => {
-            if role == "assistant" {
-                Ok(vec![ResponsesContentPart::OutputText { text: s.clone() }])
-            } else {
-                Ok(vec![ResponsesContentPart::InputText { text: s.clone() }])
-            }
-        }
-        PromptAstSimple::Media(media) => media.read_content(|c| responses_media_part(media, c)),
-        PromptAstSimple::Multiple(multiple) => {
-            let mut parts = Vec::new();
-            for item in multiple {
-                parts.extend(responses_content_parts(item, role)?);
-            }
-            Ok(parts)
-        }
-    }
-}
-
-/// Converts a media value into Responses API content parts (`input_image`, `input_audio`, `input_file`).
-fn responses_media_part(
-    media: &baml_builtins::MediaValue,
-    content: &baml_builtins::MediaContent,
-) -> Result<Vec<ResponsesContentPart>, BuildRequestError> {
-    use baml_base::MediaKind;
-    use baml_builtins::MediaContent;
-
-    match media.kind {
-        MediaKind::Image | MediaKind::Generic => {
-            let image_url = match content {
-                MediaContent::Url { url, .. } => url.clone(),
-                MediaContent::Base64 { base64_data, .. } => {
-                    format!(
-                        "data:{};base64,{}",
-                        media.mime_type.as_deref().unwrap_or("image/png"),
-                        base64_data
-                    )
-                }
-                MediaContent::File { .. } => {
-                    unreachable!("image file should have been resolved before request building")
-                }
-            };
-            Ok(vec![ResponsesContentPart::InputImage {
-                detail: Some("auto".to_string()),
-                image_url,
-            }])
-        }
-        MediaKind::Audio => match content {
-            MediaContent::Base64 { base64_data, .. } => {
-                let format = media
-                    .mime_type
-                    .as_deref()
-                    .and_then(|m| m.strip_prefix("audio/"))
-                    .map(|ext| if ext == "mpeg" { "mp3" } else { ext })
-                    .unwrap_or("mp3");
-                Ok(vec![ResponsesContentPart::InputAudio {
-                    input_audio: InputAudio {
-                        data: base64_data.clone(),
-                        format: format.to_string(),
-                    },
-                }])
-            }
-            MediaContent::Url { .. } => Err(BuildRequestError::UnsupportedMedia(
-                "audio URL is not supported on OpenAI Responses API; use base64-encoded audio instead".into(),
-            )),
-            MediaContent::File { .. } => {
-                unreachable!("audio file should have been resolved before request building")
-            }
-        },
-        MediaKind::Pdf => match content {
-            MediaContent::Url { url, .. } => Ok(vec![ResponsesContentPart::InputFile {
-                file_url: Some(url.clone()),
-                filename: Some("document.pdf".to_string()),
-                file_data: None,
-                file_id: None,
-            }]),
-            MediaContent::Base64 { base64_data, .. } => {
-                let data_url = format!(
-                    "data:{};base64,{}",
-                    media.mime_type.as_deref().unwrap_or("application/pdf"),
-                    base64_data
-                );
-                Ok(vec![ResponsesContentPart::InputFile {
-                    file_data: Some(data_url),
-                    filename: Some("document.pdf".to_string()),
-                    file_url: None,
-                    file_id: None,
-                }])
-            }
-            MediaContent::File { .. } => {
-                unreachable!("PDF file should have been resolved before request building")
-            }
-        },
-        MediaKind::Video => Err(BuildRequestError::UnsupportedMedia(
-            "video input is not supported on OpenAI Responses API".into(),
         )),
     }
 }
@@ -970,212 +746,5 @@ mod tests {
         let result = build_request(&client, msg("user", "hi")).unwrap();
         let body = parse_body(&result.body);
         assert!(body.get("model").is_none());
-    }
-
-    // ========================================================================
-    // Responses API: media content parts
-    // ========================================================================
-
-    #[test]
-    fn responses_image_url() {
-        let media = make_media(
-            MediaKind::Image,
-            MediaContent::Url {
-                url: "https://example.com/cat.png".into(),
-                base64_data: None,
-            },
-            Some("image/png"),
-        );
-        let parts = media
-            .read_content(|c| responses_media_part(&media, c))
-            .unwrap();
-        let json = serde_json::to_value(&parts[0]).unwrap();
-        assert_eq!(
-            json,
-            serde_json::json!({"type": "input_image", "detail": "auto", "image_url": "https://example.com/cat.png"})
-        );
-    }
-
-    #[test]
-    fn responses_image_base64() {
-        let media = make_media(
-            MediaKind::Image,
-            MediaContent::Base64 {
-                base64_data: "iVBORw0KGgo=".into(),
-            },
-            Some("image/png"),
-        );
-        let parts = media
-            .read_content(|c| responses_media_part(&media, c))
-            .unwrap();
-        let json = serde_json::to_value(&parts[0]).unwrap();
-        assert_eq!(json["type"], "input_image");
-        assert_eq!(json["image_url"], "data:image/png;base64,iVBORw0KGgo=");
-    }
-
-    #[test]
-    fn responses_audio_base64() {
-        let media = make_media(
-            MediaKind::Audio,
-            MediaContent::Base64 {
-                base64_data: "AAAA".into(),
-            },
-            Some("audio/wav"),
-        );
-        let parts = media
-            .read_content(|c| responses_media_part(&media, c))
-            .unwrap();
-        let json = serde_json::to_value(&parts[0]).unwrap();
-        assert_eq!(
-            json,
-            serde_json::json!({"type": "input_audio", "input_audio": {"data": "AAAA", "format": "wav"}})
-        );
-    }
-
-    #[test]
-    fn responses_pdf_url() {
-        let media = make_media(
-            MediaKind::Pdf,
-            MediaContent::Url {
-                url: "https://example.com/doc.pdf".into(),
-                base64_data: None,
-            },
-            Some("application/pdf"),
-        );
-        let parts = media
-            .read_content(|c| responses_media_part(&media, c))
-            .unwrap();
-        let json = serde_json::to_value(&parts[0]).unwrap();
-        assert_eq!(
-            json,
-            serde_json::json!({"type": "input_file", "file_url": "https://example.com/doc.pdf", "filename": "document.pdf"})
-        );
-    }
-
-    #[test]
-    fn responses_pdf_base64() {
-        let media = make_media(
-            MediaKind::Pdf,
-            MediaContent::Base64 {
-                base64_data: "JVBERi0=".into(),
-            },
-            Some("application/pdf"),
-        );
-        let parts = media
-            .read_content(|c| responses_media_part(&media, c))
-            .unwrap();
-        let json = serde_json::to_value(&parts[0]).unwrap();
-        assert_eq!(json["type"], "input_file");
-        assert_eq!(json["file_data"], "data:application/pdf;base64,JVBERi0=");
-        assert_eq!(json["filename"], "document.pdf");
-    }
-
-    #[test]
-    fn responses_video_unsupported() {
-        let media = make_media(
-            MediaKind::Video,
-            MediaContent::Url {
-                url: "https://example.com/clip.mp4".into(),
-                base64_data: None,
-            },
-            Some("video/mp4"),
-        );
-        let err = media
-            .read_content(|c| responses_media_part(&media, c))
-            .unwrap_err();
-        assert!(matches!(err, BuildRequestError::UnsupportedMedia(_)));
-    }
-
-    // ========================================================================
-    // Responses API: message building
-    // ========================================================================
-
-    #[test]
-    fn responses_assistant_uses_output_text() {
-        let parts = responses_content_parts(
-            &baml_builtins::PromptAstSimple::String("hi".into()),
-            "assistant",
-        )
-        .unwrap();
-        let json = serde_json::to_value(&parts[0]).unwrap();
-        assert_eq!(json["type"], "output_text");
-    }
-
-    #[test]
-    fn responses_user_uses_input_text() {
-        let parts =
-            responses_content_parts(&baml_builtins::PromptAstSimple::String("hi".into()), "user")
-                .unwrap();
-        let json = serde_json::to_value(&parts[0]).unwrap();
-        assert_eq!(json["type"], "input_text");
-    }
-
-    #[test]
-    fn responses_three_role_conversation() {
-        let prompt = Arc::new(PromptAst::Vec(vec![
-            msg("system", "You are a helpful assistant."),
-            msg("user", "What is 2+2?"),
-            msg("assistant", "4"),
-        ]));
-        let messages = prompt_to_responses_input(&prompt, "user").unwrap();
-        assert_eq!(messages.len(), 3);
-        assert_eq!(messages[0].role, "system");
-        assert_eq!(messages[1].role, "user");
-        assert_eq!(messages[2].role, "assistant");
-        // assistant content should use output_text
-        let json = serde_json::to_value(&messages[2].content[0]).unwrap();
-        assert_eq!(json["type"], "output_text");
-        assert_eq!(json["text"], "4");
-        // system/user content should use input_text
-        let json = serde_json::to_value(&messages[0].content[0]).unwrap();
-        assert_eq!(json["type"], "input_text");
-    }
-
-    #[test]
-    fn responses_multi_turn_conversation() {
-        let prompt = Arc::new(PromptAst::Vec(vec![
-            msg("system", "Be concise."),
-            msg("user", "Hello"),
-            msg("assistant", "Hi!"),
-            msg("user", "How are you?"),
-            msg("assistant", "Good, thanks!"),
-            msg("user", "Goodbye"),
-        ]));
-        let messages = prompt_to_responses_input(&prompt, "user").unwrap();
-        assert_eq!(messages.len(), 6);
-        assert_eq!(messages[0].role, "system");
-        assert_eq!(messages[1].role, "user");
-        assert_eq!(messages[2].role, "assistant");
-        assert_eq!(messages[3].role, "user");
-        assert_eq!(messages[4].role, "assistant");
-        assert_eq!(messages[5].role, "user");
-        // Verify assistant messages use output_text, others use input_text
-        let json2 = serde_json::to_value(&messages[2].content[0]).unwrap();
-        assert_eq!(json2["type"], "output_text");
-        let json4 = serde_json::to_value(&messages[4].content[0]).unwrap();
-        assert_eq!(json4["type"], "output_text");
-        let json5 = serde_json::to_value(&messages[5].content[0]).unwrap();
-        assert_eq!(json5["type"], "input_text");
-        assert_eq!(json5["text"], "Goodbye");
-    }
-
-    #[test]
-    fn responses_url_default() {
-        let client = make_client("openai-responses", vec![]);
-        let url = OpenAiResponsesBuilder.build_url(&client).unwrap();
-        assert_eq!(url, "https://api.openai.com/v1/responses");
-    }
-
-    #[test]
-    fn responses_url_custom_base() {
-        let client = make_client(
-            "openai-responses",
-            vec![(
-                "base_url",
-                BexExternalValue::String("https://custom.api.com/v1".into()),
-            )],
-        );
-        let url = OpenAiResponsesBuilder.build_url(&client).unwrap();
-        assert_eq!(url, "https://custom.api.com/v1/responses");
     }
 }
