+++ baml_language/crates/sys_llm/src/build_request/openai/mod.rs
@@ -1,0 +1,10 @@
+//! OpenAI-format HTTP request builders.
+//!
+//! Supports: `OpenAi`, `OpenAiGeneric`, `AzureOpenAi`, Ollama, `OpenRouter`,
+//! and `OpenAiResponses` (Responses API).
+
+mod chat_completions;
+mod responses;
+
+pub(crate) use chat_completions::OpenAiBuilder;
+pub(crate) use responses::OpenAiResponsesBuilder;
