+++ baml_language/crates/sys_llm/src/specialize_prompt/mod.rs
@@ -9,9 +9,19 @@ mod transformations;
 
 use std::str::FromStr;
 
+use bex_external_types::BexExternalValue;
 use bex_heap::builtin_types::owned::LlmPrimitiveClient;
 
 use crate::{LlmProvider, ModelFeatures};
+
+/// Check if the model name indicates an o1-family model, which does not
+/// support the system role.
+fn uses_o1_model(client: &LlmPrimitiveClient) -> bool {
+    match client.options.get("model") {
+        Some(BexExternalValue::String(m)) => m == "o1" || m.starts_with("o1-"),
+        _ => false,
+    }
+}
 
 /// Apply prompt specialization given already-extracted owned types.
 /// Specialize a prompt for a specific provider.
@@ -26,7 +36,10 @@ pub(crate) fn specialize_prompt_from_owned(
 ) -> bex_vm_types::PromptAst {
     let provider = LlmProvider::from_str(&client.provider).unwrap_or(LlmProvider::OpenAiGeneric);
 
-    let system_role_allowed = client.allowed_roles.iter().any(|r| r == "system");
+    // If the user explicitly set allowed_roles, respect that. Otherwise,
+    // detect o1-family models at runtime and disallow the system role.
+    let system_role_allowed =
+        client.allowed_roles.iter().any(|r| r == "system") && !uses_o1_model(client);
 
     let features = ModelFeatures::for_provider(provider, &client.options);
     let prompt = transformations::merge_adjacent_roles(prompt);
